feat: (PEAP-37) Protect the unauthorized routes.

+++ routes/api.php
@@ -9,19 +9,19 @@ use Illuminate\Support\Facades\Route;
 
 
 Route::middleware("jwt.auth")->group(function(){
-    Route::get("/statistics", [DashboardController::class, "index"]);
+    Route::get("/statistics", [DashboardController::class, "index"])->middleware("admin");
 
     Route::prefix("plants")->group(function () {
         Route::get("/", [PlantController::class, "index"]);
-        Route::post("/", [PlantController::class, "store"]);
+        Route::post("/", [PlantController::class, "store"])->middleware("admin");
         Route::get("/{id}", [PlantController::class, "show"]);
-        Route::put("/{id}", [PlantController::class, "update"]);
-        Route::delete("/{id}", [PlantController::class, "destroy"]);
+        Route::put("/{id}", [PlantController::class, "update"])->middleware("admin");
+        Route::delete("/{id}", [PlantController::class, "destroy"])->middleware("admin");
     });
 
     Route::prefix("orders")->group(function () {
         Route::get("/", [OrderController::class, "index"]);
-        Route::post("/", [OrderController::class, "store"]);
+        Route::post("/", [OrderController::class, "store"])->middleware("client");
         Route::get("/{id}", [OrderController::class, "show"]);
         Route::put("/{id}", [OrderController::class, "update"]);
     });
@@ -32,14 +32,14 @@ Route::middleware("jwt.auth")->group(function(){
         Route::get("/{id}", [CategoryController::class, "show"]);
         Route::put("/{id}", [CategoryController::class, "update"]);
         Route::delete("/{id}", [CategoryController::class, "destroy"]);
-    });
+    })->middleware("admin");
 });
 
 Route::prefix("auth")->group(function () {
     Route::post("/signup", [AuthController::class, "signup"]);
     Route::post("/login", [AuthController::class, "login"]);
     Route::post("/logout", [AuthController::class, "logout"]);
-});
+})->middleware("jwt.guest");
 
 Route::get('/user', function () {
     return user();
